+++ airflow/providers/apache/spark/hooks/spark_jdbc.py
@@ -19,8 +19,8 @@
 #
 import os
 
-from airflow.contrib.hooks.spark_submit_hook import SparkSubmitHook
 from airflow.exceptions import AirflowException
+from airflow.providers.apache.spark.hooks.spark_submit import SparkSubmitHook
 
 
 class SparkJDBCHook(SparkSubmitHook):
