+++ scripts/ci/pylint_todo.txt
@@ -26,7 +26,7 @@
 ./airflow/providers/apache/spark/hooks/spark_jdbc.py
 ./airflow/providers/apache/spark/hooks/spark_jdbc_script.py
 ./airflow/providers/apache/spark/hooks/spark_sql.py
-./airflow/contrib/hooks/spark_submit_hook.py
+./airflow/providers/apache/spark/hooks/spark_submit.py
 ./airflow/contrib/hooks/sqoop_hook.py
 ./airflow/contrib/hooks/vertica_hook.py
 ./airflow/contrib/hooks/wasb_hook.py
